requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
distro==1.8.0
Django==4.2.7
djangorestframework
django-allauth
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
oauthlib==3.2.2
openai==1.3.5
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
django-cors-header
django-extensions
mysqlclient
django-rest-auth
social-auth-app-django

